refactor: Split strencodings file
The files should not be duplicated between the consensus and util
libraries. Instead split off the part that is required by the consensus
library into a new file that is compiled as part of the crypto library.
