+++ src/Makefile.am
@@ -561,6 +561,8 @@ crypto_libbitcoin_crypto_base_la_SOURCES = \
   crypto/cleanse.cpp \
   crypto/cleanse.h \
   crypto/common.h \
+  crypto/hex_base.cpp \
+  crypto/hex_base.h \
   crypto/hkdf_sha256_32.cpp \
   crypto/hkdf_sha256_32.h \
   crypto/hmac_sha256.cpp \
@@ -658,8 +660,6 @@ libbitcoin_consensus_a_SOURCES = \
   tinyformat.h \
   uint256.cpp \
   uint256.h \
-  util/strencodings.cpp \
-  util/strencodings.h \
   version.h
 #
