+++ src/script/script.cpp
@@ -6,10 +6,10 @@
 #include <script/script.h>
 
 #include <crypto/common.h>
+#include <crypto/hex_base.h>
 #include <hash.h>
 #include <uint256.h>
 #include <util/hash_type.h>
-#include <util/strencodings.h>
 
 #include <string>
